+++ vite.config.ts
@@ -5,6 +5,7 @@ import renderer from 'vite-plugin-electron-renderer';
 import tailwindcss from '@tailwindcss/vite';
 
 export default defineConfig({
+  base: './',
   plugins: [
     solidPlugin(),
     tailwindcss(),